refactor: migrate ApplicationLayout and related components to XML structure
- Deleted old ApplicationLayout and pages.ts files.
- Introduced new ApplicationLayout and ApplicationView components in XML structure.
- Updated routing logic in Organization and Settings components for better clarity.
- Refactored XML adapters to use the latest core components.
- Implemented error handling and loading states in RuntimeApplicationView.
- Added zod schema for page validation in the new pages.ts file.
