+++ web/package.json
@@ -25,7 +25,6 @@
     },
     "dependencies": {
         "@astryxdesign/core": "0.3.0",
-        "@astryxdesign/core-0-3": "npm:@astryxdesign/core@0.3.0",
         "@hookform/resolvers": "^5.7.1",
         "@tailwindcss/vite": "^4.3.3",
         "@tanstack/react-query": "^5.101.4",
