--- web/src/application/runtime/pages.ts
@@ -1,16 +1,0 @@
-import { z } from 'zod';
-
-export const pageSchema = z.object({
-    tab: z.string().trim().min(1),
-    path: z.string().trim().min(1),
-    name: z.string().trim().min(1).optional(),
-    icon: z.string().trim().min(1).optional(),
-    route: z.string().trim(),
-});
-
-export type RuntimePage = z.infer<typeof pageSchema>;
-
-/** Returns true when a page route contains dynamic path segments. */
-export function pageRouteIsDynamic(route: string): boolean {
-    return /(?:^|\/):/.test(route);
-}
