+++ web/src/platform/routes/orgs/Settings.tsx
@@ -2,7 +2,6 @@ import { z } from 'zod';
 import { useState } from 'react';
 import { Text } from '@astryxdesign/core/Text';
 import { Stack } from '@astryxdesign/core/Stack';
-import { useQuery } from '@tanstack/react-query';
 import { Avatar } from '@astryxdesign/core/Avatar';
 import { Banner } from '@astryxdesign/core/Banner';
 import { HStack } from '@astryxdesign/core/HStack';
@@ -12,6 +11,7 @@ import { useLocation, useParams } from 'react-router';
 import { proportional } from '@astryxdesign/core/Table';
 import { TextInput } from '@astryxdesign/core/TextInput';
 import { EmptyState } from '@astryxdesign/core/EmptyState';
+import { skipToken, useQuery } from '@tanstack/react-query';
 import type { OrganizationStorageUsageResponse } from '@/lib/generated/platform-api-v1/types.gen';
 import { api } from '@/lib/api';
 import { S3 } from '@/components/svg/S3';
@@ -34,8 +34,6 @@ const organizationAvatarSchema = z.union([
     z.literal(''),
     z.url().refine((value) => ['http:', 'https:'].includes(new URL(value).protocol)),
 ]);
-const organizationSettingsSections = ['applications', 'database', 'storage', 'invitations', 'members'];
-const disabledApiQueryKey = ['api', 'disabled'] as const;
 
 /** Renders the organization owning a database or storage resource. */
 function OrganizationOwner({ avatar, name }: { avatar: string; name: string }) {
@@ -74,46 +72,39 @@ function OrganizationSettings() {
     const avatar = editedAvatar ?? organizationAvatar;
     const hasOrganizationApplicationAccess = hasMinimumRole(organizationRole, 'maintain');
     const hashSection = location.hash.slice(1);
-    const section = organizationSettingsSections.includes(hashSection) ? hashSection : 'organization';
     const databasePath =
-        section === 'database' && organizationId ? `/api/v1/organizations/${organizationId}/database` : null;
+        hashSection === 'database' && organizationId ? `/api/v1/organizations/${organizationId}/database` : null;
     const {
         data: databaseUsage,
         error: databaseError,
         isLoading: isDatabaseLoading,
     } = useQuery({
-        enabled: databasePath !== null,
-        queryKey: databasePath ? ['api', databasePath] : disabledApiQueryKey,
-        queryFn: async ({ signal }) => {
-            if (databasePath === null) {
-                throw new Error('Database path is unavailable');
-            }
-
-            return z
-                .int()
-                .gte(0)
-                .nullable()
-                .parse(await api(databasePath, { signal }).json());
-        },
+        queryKey: ['api', databasePath],
+        queryFn:
+            databasePath === null
+                ? skipToken
+                : async ({ signal }) =>
+                      z
+                          .int()
+                          .gte(0)
+                          .nullable()
+                          .parse(await api(databasePath, { signal }).json()),
         retry: false,
     });
     const databaseResourceError = error ?? databaseError;
     const storagePath =
-        section === 'storage' && organizationId ? `/api/v1/organizations/${organizationId}/storage` : null;
+        hashSection === 'storage' && organizationId ? `/api/v1/organizations/${organizationId}/storage` : null;
     const {
         data: storageUsage,
         error: storageError,
         isLoading: isStorageLoading,
     } = useQuery({
-        enabled: storagePath !== null,
-        queryKey: storagePath ? ['api', storagePath] : disabledApiQueryKey,
-        queryFn: async ({ signal }) => {
-            if (storagePath === null) {
-                throw new Error('Storage path is unavailable');
-            }
-
-            return zOrganizationStorageUsageResponse.nullable().parse(await api(storagePath, { signal }).json());
-        },
+        queryKey: ['api', storagePath],
+        queryFn:
+            storagePath === null
+                ? skipToken
+                : async ({ signal }) =>
+                      zOrganizationStorageUsageResponse.nullable().parse(await api(storagePath, { signal }).json()),
         retry: false,
     });
     /** Saves the Organization avatar URL when focus leaves the setting. */
