+++ web/src/xml/adapters/Dialog.tsx
@@ -1,8 +1,8 @@
 import { createContext } from 'react';
-import { Stack } from '@astryxdesign/core-0-3/Stack';
-import { Button } from '@astryxdesign/core-0-3/Button';
-import { Layout, LayoutContent } from '@astryxdesign/core-0-3/Layout';
-import { Dialog as AstryxDialog, DialogHeader } from '@astryxdesign/core-0-3/Dialog';
+import { Stack } from '@astryxdesign/core/Stack';
+import { Button } from '@astryxdesign/core/Button';
+import { Layout, LayoutContent } from '@astryxdesign/core/Layout';
+import { Dialog as AstryxDialog, DialogHeader } from '@astryxdesign/core/Dialog';
 import type { Props } from '../types';
 import { renderNode } from '../core/node';
 import { useXmlRuntime } from '../core/context';
